feat(database): add Warehouse building and adjust resource caps
- Added Warehouse building to all factions with corresponding levels.
- Adjusted default resource caps (food, wood, stone, gold) to 0.
- Improved formatting for better readability in migration files.

+++ src/game/mod.rs
@@ -4,4 +4,4 @@ pub mod res_gen_subroutine;
 pub mod resource_service;
 
 /// Server tick rate in seconds
-pub static TICK_RATE: u8 = 60;
+pub static TICK_RATE: i32 = 60;
